remove ganache.py's dependency on constants.py, by moving confFileValue() from constants.py to its own utility file configutil.py
