+++ web3tools/test/test_configutil.py
@@ -1,0 +1,4 @@
+from util import configutil
+
+def test1():
+    assert isinstance(configutil.CONF_FILE_PATH, str)
